05.06.2025 - "Finalização portal RH"

+++ Frameworks/Angular/sa_portal_rh/src/app/views/painel-curriculos/painel-curriculos.component.ts
@@ -8,7 +8,7 @@ import { CurriculoService } from 'src/app/services/curriculo.service';
   styleUrls: ['./painel-curriculos.component.scss']
 })
 export class PainelCurriculosComponent {
-  public curriculo: Curriculo = new Curriculo("", "", 0, "", "", [], "", "", [], [], [], []);
+  public curriculo: Curriculo = new Curriculo(0,"", "", 0, "", "", [], "", "", [], [], [], [], "");
 
   public curriculos: Curriculo[] = [];
 
@@ -22,6 +22,7 @@ export class PainelCurriculosComponent {
     this._curriculoService.getCurriculos().subscribe((retornaCurriculo) => {
       this.curriculos = retornaCurriculo.map((item) => {
         return new Curriculo(
+          item.id,
           item.cpf,
           item.nome,
           item.idade,
@@ -34,6 +35,7 @@ export class PainelCurriculosComponent {
           item.qualificacoes,
           item.habilidades,
           item.idiomas,
+          item.sobre,
         );
       });
     });
@@ -44,29 +46,66 @@ export class PainelCurriculosComponent {
   }
 
   cadastrarCurriculo() {
-    this._curriculoService.cadastrarCurriculo(this.curriculo).subscribe(() => {
-      this.curriculo = new Curriculo("", "", 0, "", "", [], "", "", [], [], [], []);
-      this.listarCurriculos();
-      alert("Curriculo cadastrada com sucesso");
-    }, (err) => {
-      console.error("Exception: ", err);
-    });
-  }
+    if (!this.curriculo.cpf || !this.curriculo.nome || !this.curriculo.idade || !this.curriculo.email || !this.curriculo.contato) {
+      alert("Preencha todos os campos obrigatórios.");
+      return;
+    }
+    if (this.curriculo.idade < 18) {
+      alert("A idade mínima para cadastro é 18 anos.");
+      return;
+    }
 
-atualizarCurriculo(id:any){
-      this._curriculoService.atualizarCurriculo(id, this.curriculo).subscribe(() => {
-      this.curriculo = new Curriculo("", "", 0, "", "", [], "", "", [], [], [], []);
-      this.listarCurriculos();
-      alert("Curriculo atualizada com sucesso");
-    }, (err) => {
-      console.error("Exception: ", err);
+    this.curriculo.interesses!.toString().split(",").forEach((interesse) => {
+      if (interesse.trim() === "") {
+        alert("Interesses não podem estar vazios.");
+        throw new Error("Interesses não podem estar vazios.");
+      }
     });
-  }
+
+    this.curriculo.interesses = this.curriculo.interesses!.toString().split(",").map((interesse) => interesse.trim());
+    this.curriculo.formacao = this.curriculo.formacao!.toString().split(",").map((formacao) => formacao.trim());
+    this.curriculo.qualificacoes = this.curriculo.qualificacoes!.toString().split(",").map((qualificacao) => qualificacao.trim());
+    this.curriculo.habilidades = this.curriculo.habilidades!.toString().split(",").map((habilidade) => habilidade.trim());
+    this.curriculo.idiomas = this.curriculo.idiomas!.toString().split(",").map((idioma) => idioma.trim());
+    this.curriculo.fotoPerfil = "foto-perfil";
+  this._curriculoService.cadastrarCurriculo(this.curriculo).subscribe(() => {
+    this.curriculo = new Curriculo(0, "", "", 0, "", "", [], "", "", [], [], [], [], "");
+    this.listarCurriculos();
+    alert("Currículo cadastrado com sucesso");
+  }, (err) => {
+    console.error("Exception: ", err);
+  });
+    
+    
+  };
+  
+
+atualizarCurriculo(id: any) {
+    this._curriculoService.atualizarCurriculo(id, this.curriculo).subscribe({
+      next: () => {
+        this.curriculo = new Curriculo(0,"","", 0, "", "", [], "", "", [], [], [], [], "");
+        this.listarCurriculos();
+        alert("Curriculo atualizada com sucesso");
+      },
+      error: (err) => {
+        console.error("Exception: ", err);
+      }
+    });
+}
 
 deletarCurriculo(id:any){
+  if (!id) {
+    alert("ID do currículo não fornecido.");
+    return;
+  }
+
+  if (!confirm("Tem certeza que deseja excluir este currículo?\n CPF: " + this.curriculo.cpf + "\n Id: " + this.curriculo.id)) {
+    return;
+  }
+
   this._curriculoService.removerCurriculo(id).subscribe(()=>{
     this.listarCurriculos();
-    alert("Curriculo excluída com sucesso");
+    alert("Curriculo excluído com sucesso");
   }, (err) =>{
     console.error("Exception: ",err)
   })
